+++ google/cloud/cloudcontrolspartner/quickstart/quickstart.cc
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 //! [all]
-#include "google/cloud/cloudcontrolspartner/v1/ EDIT HERE _client.h"
+#include "google/cloud/cloudcontrolspartner/v1/cloud_controls_partner_core_client.h"
 #include "google/cloud/location.h"
 #include <iostream>
 
@@ -26,10 +26,10 @@ int main(int argc, char* argv[]) try {
   auto const location = google::cloud::Location(argv[1], argv[2]);
 
   namespace cloudcontrolspartner = ::google::cloud::cloudcontrolspartner_v1;
-  auto client = cloudcontrolspartner::ServiceClient(
-      cloudcontrolspartner::MakeServiceConnection());  // EDIT HERE
+  auto client = cloudcontrolspartner::CloudControlsPartnerCoreClient(
+      cloudcontrolspartner::MakeCloudControlsPartnerCoreConnection());
 
-  for (auto r : client.List /*EDIT HERE*/ (location.FullName())) {
+  for (auto r : client.ListCustomers(location.FullName())) {
     if (!r) throw std::move(r).status();
     std::cout << r->DebugString() << "\n";
   }